# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateWebhookRequest:
    properties:
      events:
        example:
        - message
        - ack
        - call.received
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        example:
          Authorization: Bearer token123
        type: object
      name:
        example: N8N Webhook
        type: string
      retry_count:
        example: 3
        type: integer
      secret:
        example: my-secret-key
        type: string
      timeout_seconds:
        example: 30
        type: integer
      url:
        example: https://n8n.example.com/webhook/waha-events
        type: string
    required:
    - events
    - name
    - url
    type: object
...
  handlers.UpdateWebhookRequest:
    properties:
      active:
        type: boolean
      events:
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      retry_count:
        type: integer
      secret:
        type: string
      timeout_seconds:
        type: integer
      url:
        type: string
    type: object
...
info:
  contact:
    email: support@ventros.com
    name: API Support
  description: API para gerenciamento de CRM com eventos e workflows
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ventros CRM API
  version: "1.0"
paths:
...
  /api/v1/webhook-subscriptions:
    get:
      description: Lista todas as inscrições de webhooks
      parameters:
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of webhooks
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List webhook subscriptions
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Cria uma nova inscrição de webhook para receber eventos do WAHA
      parameters:
      - description: Webhook subscription data
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Webhook created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create webhook subscription
      tags:
      - webhooks
  /api/v1/webhook-subscriptions/{id}:
    delete:
      description: Remove uma inscrição de webhook
      parameters:
      - description: Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Webhook deleted successfully
        "404":
          description: Webhook not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete webhook subscription
      tags:
      - webhooks
    get:
      description: Obtém detalhes de uma inscrição de webhook
      parameters:
      - description: Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Webhook not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get webhook subscription
      tags:
      - webhooks
    put:
      consumes:
      - application/json
      description: Atualiza uma inscrição de webhook
      parameters:
      - description: Webhook ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Webhook update data
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Webhook not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update webhook subscription
      tags:
      - webhooks
...
  /api/v1/webhooks/{webhook_id}:
    post:
      consumes:
      - application/json
      description: Recebe eventos de webhook do WAHA usando ID único do webhook (padrão
        indústria)
      parameters:
      - description: Webhook ID único
        in: path
        name: webhook_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event queued
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Webhook not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Receive WAHA webhook
      tags:
      - webhooks
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication for service-to-service communication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"